powerpc: gamecube/wii: udbg support via framebuffer
Add support for using the Video Interface (VI) framebuffer of the
Nintendo GameCube and Wii video game consoles as a udbg console.

As only one udbg console (gcnvi or usbgecko) can be used at a given
point, a new choice option is introduced to select between the two available
options.

Signed-off-by: Albert Herranz <albert_herranz@yahoo.es>