wii: platform restart rework
Use the new IOS reload helpers in the platform restart code
and kexec path.

A platform restart tries now to load The Homebrew Channel and falls back
to a normal STM restart if the former fails.
On the other hand, the kexec reboot path reloads IOS before the final
kexec phase to ensure that the kexec'ed image will get IOS in a known
clean state.

Signed-off-by: Albert Herranz <albert_herranz@yahoo.es>