wii: mini flavour: use mmcblk0p2 as root device and wait
On the traditional IOS-dependent kernel, rvlsda is the block device
associated to the front SD card slot.
But on a 'mini' kernel, the front SD card slot is driven by a SDHCI-based
host controller driver integrated into the Linux MMC framework.
In this case, mmcblk0 is the block device to use.

Our defacto standard is still to look for the root filesystem on the second
partition of the SD card inserted into the front SD card slot.

Signed-off-by: Albert Herranz <albert_herranz@yahoo.es>