wii: ohci-mipc: add control list quirk
The OHCI USB host controllers on the Nintendo Wii video game console
stop working when new TDs are added to a scheduled control ED after a
transfer has has taken place on it.

Add a new quirk and a workaround function to deal with this situation.
Before scheduling any new control TD, we make the controller happy by
always loading a special control ED with a single dummy TD and letting
the controller attempt the transfer.
The controller won't do anything with it, as the special ED has no TDs,
but it will keep the controller from failing on the next transfer.

Signed-off-by: Albert Herranz <albert_herranz@yahoo.es>