wii: ohci-mipc: add bulk list quirk
There seem to be issues too with the bulk list processing on the
OHCI controller found in the Nintendo Wii video game console.
The exact problem remains still unidentified, but adding a small
delay at the proper place seems to workaround it.

As an example, without this quirk the wiimote controller stops
responding after a few seconds because one of its bulk endpoint
descriptors gets stuck.

Signed-off-by: Albert Herranz <albert_herranz@yahoo.es>