sdio: recognize io card without powercycle
SDIO Simplified Specification (at least V2.00) states that it is strongly
recommended that the host executes either a power reset or issues a CMD52
(I/O Reset) to re-initialize an I/O card.
Additionally, the CMD52 must be issued first because it cannot be issued
after a CMD0.

With this patch the Nintendo Wii SDIO-based WLAN card is detected after a
system reset, without requiring a complete system powercycle.

Signed-off-by: Albert Herranz <albert_herranz@yahoo.es>